refactor: extract Provider.Verification
Move verification-document submission plus the admin review workflow
(approve/reject, listing, signed-URL preview) and all supporting helpers
onto Provider.Verification; facade delegates. verify_provider/unverify_provider
stay behind for the Profiles slice (they mutate ProviderProfile).
